+++ CHANGELOG.md
@@ -10,17 +10,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
  -  Ability to create HWMs
  -  Dropdown functionality to state and event filters on map page
+ -  Added ability to delete HWMs
+ -  Ability to create Reference Datums
+ -  Added ability to delete reference datums
+ -  Added ability to deploy a sensor
+ -  Added ability to delete sensors
  
 ### Changed
 
  -  Clear filters button styling to link
  -  Increased details modal widths
  -  Improved responsiveness of tables in details modals
+ -  Autopopulate source and agency when adding file in site edit form
 
  
  ### Fixed
  
- -
+ -  Time reset in sensor edit modal when Cancel Edits is clicked
 
 ## [v0.7.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.7.0) - 2022-01-17